backend/requirements.txt
fastapi==0.115.2
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==2.9.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-dotenv==1.0.1
transformers==4.36.2
rapidfuzz==3.7.1
scikit-learn==1.4.2
pandas==2.2.2
pytz==2024.1